Rename and rearrange classes in spring-web module
(cherry picked from commit 5a198020bf9aaad9b4933bbe63d60703c85be630)
