--- spring/spring-web/src/test/java/io/quarkus/ts/spring/web/reactive/MariaDBUtils.java
@@ -1,6 +1,0 @@
-package io.quarkus.ts.spring.web.reactive;
-
-public class MariaDBUtils {
-    public static final int PORT = 3306;
-    public static final String START_LOG = "socket: '/run/mysqld/mysqld.sock'  port: " + PORT;
-}
