+++ spring/spring-web/src/test/java/io/quarkus/ts/spring/web/reactive/boostrap/OpenShiftSpringWebQuteIT.java
@@ -1,18 +1,19 @@
 package io.quarkus.ts.spring.web.reactive.boostrap;
 
+import static io.quarkus.ts.spring.web.reactive.MariaDBConstants.IMAGE_105;
+import static io.quarkus.ts.spring.web.reactive.MariaDBConstants.PORT;
+import static io.quarkus.ts.spring.web.reactive.MariaDBConstants.START_LOG_105;
+
 import io.quarkus.test.bootstrap.MariaDbService;
 import io.quarkus.test.bootstrap.RestService;
 import io.quarkus.test.scenarios.OpenShiftScenario;
 import io.quarkus.test.services.Container;
 import io.quarkus.test.services.QuarkusApplication;
-import io.quarkus.ts.spring.web.reactive.AbstractDbIT;
 
 @OpenShiftScenario
-public class OpenShiftHomePageIT extends AbstractDbIT {
+public class OpenShiftSpringWebQuteIT extends AbstractSpringWebQuteIT {
 
-    static final int MARIADB_PORT = 3306;
-
-    @Container(image = "${mariadb.105.image}", port = MARIADB_PORT, expectedLog = "Only MySQL server logs after this point")
+    @Container(image = IMAGE_105, port = PORT, expectedLog = START_LOG_105)
     static final MariaDbService database = new MariaDbService();
 
     @QuarkusApplication
