+++ spring/spring-web/src/test/java/io/quarkus/ts/spring/web/reactive/openapi/AbstractSpringWebOpenApiIT.java
@@ -6,43 +6,21 @@ import java.util.stream.Stream;
 
 import org.apache.http.HttpStatus;
 import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.MethodOrderer;
 import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.TestMethodOrder;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvFileSource;
 
-import io.quarkus.test.bootstrap.MariaDbService;
 import io.quarkus.test.bootstrap.RestService;
-import io.quarkus.test.scenarios.QuarkusScenario;
-import io.quarkus.test.services.Container;
-import io.quarkus.test.services.QuarkusApplication;
-import io.quarkus.ts.spring.web.reactive.AbstractDbIT;
-import io.quarkus.ts.spring.web.reactive.MariaDBUtils;
 import io.restassured.response.Response;
 
-@QuarkusScenario
-@TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-public class OpenApiIT extends AbstractDbIT {
+public abstract class AbstractSpringWebOpenApiIT {
     private static Response response;
 
-    @Container(image = "${mariadb.10.image}", port = MariaDBUtils.PORT, expectedLog = MariaDBUtils.START_LOG)
-    static final MariaDbService database = new MariaDbService();
-
-    @QuarkusApplication
-    private static final RestService app = new RestService()
-            .withProperty("quarkus.datasource.username", database.getUser())
-            .withProperty("quarkus.datasource.password", database.getPassword())
-            .withProperty("quarkus.datasource.jdbc.url", database::getJdbcUrl);
-
-    @Override
-    public RestService getApp() {
-        return app;
-    }
+    protected abstract RestService getApp();
 
     private void callOpenApiEndpoint() {
-        response = app.given().get("/q/openapi?format=JSON");
+        response = getApp().given().get("/q/openapi?format=JSON");
     }
 
     @Order(1)
@@ -75,4 +53,5 @@ public class OpenApiIT extends AbstractDbIT {
         Assertions.assertEquals(1, objectMap.keySet().size());
         Assertions.assertEquals(type, objectMap.keySet().iterator().next());
     }
+
 }
